refactor: rename user service to identity service
